backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
bcrypt==4.1.3
python-multipart==0.0.9
Pillow==10.3.0